Changed model.ppv.PPVModel.[save/load]
Problem: Saving the model has been problematic as pymc3 changed from MultiTrace
to InfrenceData as the recommended output between python 3.7 and 3.9,
and pandas changed major version making old pandas unpickalable unless
they are pickled into their own file

Solution:
 * PPVModel.save now creates a zip file with df and trace pickled by
   themselvs, it also pickles the arguments to __init__ so the object
   can be recreated, it takes a second argument, which if set to 'small'
   or 'sample' then it only saves the posterior samples, which is all
   that is needed for prediction, but for many other features such as
   the log posterior is needed which is contained in InfrenceData and
   thus saved by default (2nd arg set to 'pickle').
 * PPVModel.load has been updated to work the same way, this makes the
   model unable to load old models (they can still be loaded with
   pickle.load) which was how models used to be saved.

added tests for the new behaviour in tests/test_model.py

+++ ppv/_pandas_extension.py
@@ -230,7 +230,7 @@ class PandasDataFramePPVFeatures:
         return df
 
     def get_prior(self):
-        return self.target.astype(bool).sum() / self.target.shape[0]
+        return float(self.target.astype(bool).sum() / self.target.shape[0])
 
     @classmethod
     def _seq_to_modseq(cls, row):
